refactor: 인증을 위한 커스텀 예외 클래스 생성
