--- src/main/java/roomescape/global/exception/ErrorType.java
@@ -1,24 +1,0 @@
-package roomescape.global.exception;
-
-import org.springframework.http.HttpStatus;
-
-public enum ErrorType {
-    ACCESS_FORBIDDEN(HttpStatus.FORBIDDEN, "접근 권한이 없습니다."),
-    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "로그인이 필요합니다.");
-
-    private final HttpStatus httpStatus;
-    private final String errorMessage;
-
-    ErrorType(final HttpStatus httpStatus, final String errorMessage) {
-        this.httpStatus = httpStatus;
-        this.errorMessage = errorMessage;
-    }
-
-    public HttpStatus getHttpStatus() {
-        return httpStatus;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-}
